+++ src/Transfer/Subscription/DeleteSubscriber.cs
@@ -7,6 +7,6 @@ namespace Transfer.Subscription
     public static class DeleteSubscriber
     {
         public record Request
-            (string Url, int Id, string Token) : IRequest<OneOf<Unit, ValidationError, DatabaseError>>;
+            (string Url, long Id, string Token) : IRequest<OneOf<Unit, ValidationError, DatabaseError>>;
     }
 }
